requirements.txt
# Web dashboard
streamlit>=1.30.0

# Text extraction
PyMuPDF>=1.23.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0

# TTS engines
kokoro-onnx>=0.4.0
soundfile>=0.12.0

# Audio processing
pydub>=0.25.1

# Configuration
pyyaml>=6.0
